requirements.txt
Django==1.8.7
Markdown==2.6.5
djangorestframework==3.3.1
